pyproject.toml
[project]
name = "evals"
version = "0.1.0"
description = "Simple LLM product evals demo"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai",
    "marimo>=0.18.1",
    "matplotlib>=3.10.7",
    "openai>=1.0",
    "python-dotenv",
]
